feat(domainStats): Support querying domain stat fields (#6664)
* refactor Company to SDL

* support activeOrganizationCount and meetingCount

* add after args to aggregate fields

* feat(insights): add insights feature flag

* refactor getTeamsByOrgIds

* clean up old flags
