+++ packages/server/graphql/public/typeDefs/Company.graphql
@@ -1,0 +1,65 @@
+"""
+A grouping of organizations. Automatically grouped by top level domain of each
+"""
+type Company {
+  """
+  the top level domain
+  """
+  id: ID!
+
+  """
+  the number of active teams across all organizations
+  """
+  activeTeamCount(
+    """
+    Only include teams that have met since this time.
+    Defaults to 30 days ago
+    """
+    after: DateTime
+  ): Int!
+
+  """
+  the number of active users across all organizations
+  """
+  activeUserCount(
+    """
+    Only include users that joined after this time
+    """
+    after: DateTime
+  ): Int!
+
+  """
+  the number of active workspaces used by the company
+  """
+  activeOrganizationCount: Int!
+
+  """
+  the last time any team in the organization started a meeting, null if no meetings were ever run
+  """
+  lastMetAt: DateTime
+
+  """
+  the total number of meetings started across all teams on all organizations
+  """
+  meetingCount(after: DateTime): Int!
+
+  """
+  the longest monthly streak for meeting at least once per month for any team in the company
+  """
+  monthlyTeamStreakMax: Int!
+
+  """
+  Get the list of all organizations that belong to the company
+  """
+  organizations: [Organization!]!
+
+  """
+  The highest tier for any organization within the company
+  """
+  tier: TierEnum!
+
+  """
+  the total number of users across all organizations
+  """
+  userCount: Int!
+}
