+++ packages/server/graphql/public/types/User.ts
@@ -1,0 +1,17 @@
+import {isSuperUser} from '../../../utils/authorization'
+import getDomainFromEmail from '../../../utils/getDomainFromEmail'
+import isCompanyDomain from '../../../utils/isCompanyDomain'
+import {UserResolvers} from '../resolverTypes'
+
+const User: UserResolvers = {
+  company: async ({email}, _args, {authToken}) => {
+    const domain = getDomainFromEmail(email)
+    if (!domain || !isCompanyDomain(domain) || !isSuperUser(authToken)) return null
+    return {id: domain}
+  },
+  featureFlags: ({featureFlags}) => {
+    return Object.fromEntries(featureFlags.map((flag) => [flag as any, true]))
+  }
+}
+
+export default User
