+++ packages/server/graphql/rootMutation.ts
@@ -4,7 +4,6 @@ import acceptTeamInvitation from './mutations/acceptTeamInvitation'
 import addAgendaItem from './mutations/addAgendaItem'
 import addAtlassianAuth from './mutations/addAtlassianAuth'
 import addComment from './mutations/addComment'
-import addFeatureFlag from './mutations/addFeatureFlag'
 import addGitHubAuth from './mutations/addGitHubAuth'
 import addIntegrationProvider from './mutations/addIntegrationProvider'
 import addMissingJiraField from './mutations/addMissingJiraField'
@@ -158,7 +157,6 @@ export default new GraphQLObjectType<any, GQLContext>({
       addReflectTemplate,
       addReflectTemplatePrompt,
       addSlackAuth,
-      addFeatureFlag,
       addGitHubAuth,
       addOrg,
       addTeam,
